+++ demos/client-example-server/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @prosopo/client-example-server
 
+## 2.7.137
+### Patch Changes
+
+- Updated dependencies [ae475a5]
+  - @prosopo/types@5.2.2
+  - @prosopo/database@4.0.11
+  - @prosopo/keyring@2.9.73
+  - @prosopo/server@2.10.54
+
+## 2.7.136
+### Patch Changes
+
+- Updated dependencies [35f640f]
+  - @prosopo/types@5.2.1
+  - @prosopo/database@4.0.10
+  - @prosopo/keyring@2.9.72
+  - @prosopo/server@2.10.53
+
+## 2.7.135
+### Patch Changes
+
+- Updated dependencies [234c737]
+  - @prosopo/types@5.2.0
+  - @prosopo/database@4.0.9
+  - @prosopo/keyring@2.9.71
+  - @prosopo/server@2.10.52
+
+## 2.7.134
+### Patch Changes
+
+- Updated dependencies [ee5d250]
+  - @prosopo/types@5.1.2
+  - @prosopo/database@4.0.8
+  - @prosopo/keyring@2.9.70
+  - @prosopo/server@2.10.51
+
+## 2.7.133
+### Patch Changes
+
+- Updated dependencies [cec44bb]
+  - @prosopo/types@5.1.1
+  - @prosopo/database@4.0.7
+  - @prosopo/keyring@2.9.69
+  - @prosopo/server@2.10.50
+
 ## 2.7.132
 ### Patch Changes